requirements.txt
amqp==5.2.0
asgiref==3.4.1
async-timeout==4.0.3
attrs==21.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.28.0
coverage==6.2
cron-descriptor==1.4.3
dj-database-url==0.5.0
Django==3.2.9
django-celery-beat==2.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
django-timezone-field==7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
exceptiongroup==1.2.2
factory-boy==3.2.1
Faker==10.0.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.3.1
kombu==5.4.0
packaging==24.1
Pillow==8.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pyrsistent==0.18.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
redis==5.0.8
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.1.0
wcwidth==0.2.13
whitenoise==5.3.0
zipp==3.6.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.7,<1.0